Unreviewed Leopard build fix attempt.
Canonical link: https://commits.webkit.org/60832@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@70281 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/rendering/InlineTextBox.cpp
@@ -430,7 +430,7 @@ void InlineTextBox::paint(PaintInfo& paintInfo, int tx, int ty)
     if (m_isVertical) {
         context->save();
         context->translate(textOrigin.x(), textOrigin.y());
-        context->rotate(deg2rad(90.));
+        context->rotate(static_cast<float>(deg2rad(90.)));
         context->translate(-textOrigin.x(), -textOrigin.y());
     }
